[TARGETING] Inject targeting.js independently from Piwik and rename targeting JS file
